ai: extract the shared ACP stdio transport
The kiro-cli provider carried its own JSON-RPC plumbing for driving an
Agent Client Protocol server over stdio: spawn, initialize, open a
session, send one prompt, collect the streamed agent text, and keep a
redacted stderr preview for error messages. None of that is specific to
kiro-cli, and other agents expose the same protocol.

Move it to a new ai::acp module behind an AcpProcess description of the
child to run. Error messages keep their exact wording through the label
field, and the session cwd, working directory and child environment
become parameters so an agent with stricter expectations can be driven
by the same transport. The returned AcpPrompt also carries the raw
session/prompt result, which some agents use to report token usage.

No functional change: kiro-cli keeps the same arguments, the same
isolated no-tool agent, the same deny-all hook and the same relative
session cwd.
